+++ backend/src/main/java/ristaurace/repository/MenuPolozkaMenuRepository.java
@@ -5,6 +5,6 @@ package ristaurace.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ristaurace.entities.MenuPolozkaMenuEntity;
 
-public interface MenuPolozkaMenuRepository extends JpaRepository<MenuPolozkaMenuEntity, Long> {
+public interface MenuPolozkaMenuRepository extends JpaRepository<MenuPolozkaMenuEntity, Integer> {
 
 }
