+++ backend/src/main/java/ristaurace/repository/MenuRepository.java
@@ -5,6 +5,6 @@ package ristaurace.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ristaurace.entities.MenuEntity;
 
-public interface MenuRepository extends JpaRepository<MenuEntity, Long> {
+public interface MenuRepository extends JpaRepository<MenuEntity, Integer> {
 
 }
