+++ backend/src/main/java/ristaurace/repository/StulRepository.java
@@ -7,6 +7,6 @@ import ristaurace.entities.StulEntity;
 
 import java.util.List;
 
-public interface StulRepository extends JpaRepository<StulEntity, Long> {
+public interface StulRepository extends JpaRepository<StulEntity, Integer> {
     List<StulEntity> findAll();
 }
